+++ Changelog.md
@@ -1,3 +1,10 @@
+## 04.10.2024
+1. Removed Forta integration.
+2. Fixed issues with Telegram markdown formatting.
+3. Improved error handling: if FF fails to send a message with Telegram markdown, it will now send it as plain text.
+4. Implemented length checks for Telegram: messages exceeding 4,096 characters will be truncated.
+5. Implemented length checks for Discord: messages exceeding 2,000 characters will be truncated.
+
 ## 24.09.2024
 1. Fix sending network alerts though telegram